requirements.txt
elevenlabs==1.11.0
python-dotenv==1.0.1
requests==2.32.3
